refactor: move rgbToHex & paginate functions to shared util folder
