+++ src/utils/__tests__/rgb-to-hex.test.ts
@@ -1,4 +1,4 @@
-import { rgbToHex } from '@/graphql/resolvers/utils'
+import { isValidHexCode, rgbToHex } from '@/utils'
 
 it('returns an array of Hex colors', async () => {
   const colors: [red: number, green: number, blue: number][] = [
@@ -14,8 +14,6 @@ it('returns an array of Hex colors', async () => {
   const result = colors.map((color) => rgbToHex(color))
   expect(result).toBeDefined()
   result.forEach((el) => {
-    //regEx to check if each string in the array is a valid hex color format
-    //https://stackoverflow.com/questions/1636350/how-to-identify-a-given-string-is-hex-color-format
-    expect(el).toMatch(/^#(?:(?:[\da-f]{3}){1,2}|(?:[\da-f]{4}){1,2})$/i)
+    expect(el).toMatch(isValidHexCode)
   })
 })
