+++ src/utils/index.ts
@@ -1,3 +1,5 @@
 export * from './capitalize'
-export * from './getColorsForExport'
+export * from './get-colors-for-export'
+export * from './paginate'
+export * from './rgb-to-hex'
 export * from './test-utils'
