+++ src/utils/paginate.ts
@@ -1,10 +1,5 @@
 import { albums } from '@prisma/client'
 
-//Helper function to covert rgb color codes to HEX color codes
-export const rgbToHex = (arr: [number, number, number]): string => {
-  return '#' + arr.map((v) => ('0' + v.toString()).slice(-2)).join('')
-}
-
 //Helper function to paginate queries
 export const paginate = (
   firstArg: number,
